+++ assets/graph_node/graphim_node.gd
@@ -2,6 +2,10 @@
 class_name GraphimNode extends DraggableObject
 
 
+## Emitida cuando se le hace clic al nodo. Útil para dibujar aristas
+signal clicked(graphim_node: GraphimNode)
+
+
 ## Sprite del nodo
 @onready var sprite: Sprite2D = $Sprite
 ## Forma de colisión del nodo
@@ -164,3 +168,15 @@ func get_data_copy() -> NodeData:
 
 
 #endregion
+
+
+# Procesa los eventos (filtrando clics) para el dibujado de aristas
+func _on_input_event(_viewport, event: InputEvent, _shape_idx) -> void:
+	# Filtra los clics izquierdos
+	if not (
+		event is InputEventMouseButton
+		and event.is_pressed()
+		and event.button_index == MOUSE_BUTTON_LEFT
+	): return
+
+	clicked.emit(self)